+++ examples/transactions.rs
@@ -6,9 +6,9 @@ use fred::prelude::*;
 #[tokio::main]
 async fn main() -> Result<(), RedisError> {
   let client = RedisClient::default();
-  let _ = client.connect();
-  client.wait_for_connect().await?;
+  client.init().await?;
 
+  // transactions are buffered in memory before calling `exec`
   let trx = client.multi();
   let result: RedisValue = trx.get("foo").await?;
   assert!(result.is_queued());
@@ -17,6 +17,8 @@ async fn main() -> Result<(), RedisError> {
   let result: RedisValue = trx.get("foo").await?;
   assert!(result.is_queued());
 
+  // automatically send `WATCH ...` before `MULTI`
+  trx.watch_before(vec!["foo", "bar"]);
   let values: (Option<String>, (), String) = trx.exec(true).await?;
   println!("Transaction results: {:?}", values);